make ConfigTestCases templated to allow extending
