+++ test/configCases/library/1-use-library/webpack.config.js
@@ -1,14 +1,11 @@
 var webpack = require("../../../../");
 var path = require("path");
-/** @type {import("../../../../").Configuration[]} */
-module.exports = [
+/** @type {function(any, any): import("../../../../").Configuration[]} */
+module.exports = (env, { testPath }) => [
 	{
 		resolve: {
 			alias: {
-				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/commonjs.js"
-				)
+				library: path.resolve(testPath, "../0-create-library/commonjs.js")
 			}
 		},
 		plugins: [
@@ -20,10 +17,7 @@ module.exports = [
 	{
 		resolve: {
 			alias: {
-				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/umd.js"
-				)
+				library: path.resolve(testPath, "../0-create-library/umd.js")
 			}
 		},
 		plugins: [
@@ -36,10 +30,7 @@ module.exports = [
 		entry: "./this-test.js",
 		resolve: {
 			alias: {
-				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/this.js"
-				)
+				library: path.resolve(testPath, "../0-create-library/this.js")
 			}
 		},
 		plugins: [
@@ -52,10 +43,7 @@ module.exports = [
 		entry: "./var-test.js",
 		resolve: {
 			alias: {
-				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/var.js"
-				)
+				library: path.resolve(testPath, "../0-create-library/var.js")
 			}
 		},
 		plugins: [
@@ -68,8 +56,8 @@ module.exports = [
 		resolve: {
 			alias: {
 				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/commonjs2-external.js"
+					testPath,
+					"../0-create-library/commonjs2-external.js"
 				),
 				external: path.resolve(__dirname, "node_modules/external.js")
 			}
@@ -85,8 +73,8 @@ module.exports = [
 		resolve: {
 			alias: {
 				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/commonjs2-split-chunks/"
+					testPath,
+					"../0-create-library/commonjs2-split-chunks/"
 				),
 				external: path.resolve(__dirname, "node_modules/external.js")
 			}
@@ -101,10 +89,7 @@ module.exports = [
 		entry: "./default-test.js",
 		resolve: {
 			alias: {
-				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/umd-default.js"
-				)
+				library: path.resolve(testPath, "../0-create-library/umd-default.js")
 			}
 		},
 		plugins: [
@@ -116,10 +101,7 @@ module.exports = [
 	{
 		resolve: {
 			alias: {
-				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/entryA.js"
-				)
+				library: path.resolve(testPath, "../0-create-library/entryA.js")
 			}
 		},
 		plugins: [
@@ -131,10 +113,7 @@ module.exports = [
 	{
 		resolve: {
 			alias: {
-				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/entryB.js"
-				)
+				library: path.resolve(testPath, "../0-create-library/entryB.js")
 			}
 		},
 		plugins: [
@@ -146,10 +125,7 @@ module.exports = [
 	{
 		resolve: {
 			alias: {
-				library: path.resolve(
-					__dirname,
-					"../../../js/config/library/0-create-library/entryC.js"
-				)
+				library: path.resolve(testPath, "../0-create-library/entryC.js")
 			}
 		},
 		plugins: [
