+++ test/configCases/plugins/lib-manifest-plugin/webpack.config.js
@@ -1,21 +1,18 @@
 var path = require("path");
 var LibManifestPlugin = require("../../../../").LibManifestPlugin;
 
-/** @type {import("../../../../").Configuration} */
-module.exports = {
+/** @type {function(any, any): import("../../../../").Configuration} */
+module.exports = (env, { testPath }) => ({
 	entry: {
 		bundle0: ["./"]
 	},
 	plugins: [
 		new LibManifestPlugin({
-			path: path.resolve(
-				__dirname,
-				"../../../js/config/plugins/lib-manifest-plugin/[name]-manifest.json"
-			),
+			path: path.resolve(testPath, "[name]-manifest.json"),
 			name: "[name]_[fullhash]"
 		})
 	],
 	node: {
 		__dirname: false
 	}
-};
+});
