+++ test/configCases/records/issue-2991/webpack.config.js
@@ -1,12 +1,9 @@
 var path = require("path");
 
-/** @type {import("../../../../").Configuration} */
-module.exports = {
+/** @type {function(any, any): import("../../../../").Configuration} */
+module.exports = (env, { testPath }) => ({
 	entry: "./test",
-	recordsOutputPath: path.resolve(
-		__dirname,
-		"../../../js/config/records/issue-2991/records.json"
-	),
+	recordsOutputPath: path.resolve(testPath, "records.json"),
 	target: "node",
 	node: {
 		__dirname: false
@@ -17,4 +14,4 @@ module.exports = {
 			pkgs: path.resolve(__dirname, "pkgs")
 		}
 	}
-};
+});
