+++ test/configCases/records/issue-7339/webpack.config.js
@@ -1,14 +1,11 @@
 var path = require("path");
 
-/** @type {import("../../../../").Configuration} */
-module.exports = {
+/** @type {function(any, any): import("../../../../").Configuration} */
+module.exports = (env, { testPath }) => ({
 	entry: "./test",
-	recordsOutputPath: path.resolve(
-		__dirname,
-		"../../../js/config/records/issue-7339/records.json"
-	),
+	recordsOutputPath: path.resolve(testPath, "records.json"),
 	target: "node",
 	node: {
 		__dirname: false
 	}
-};
+});
